# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DailyReportResponse": {
        "properties": {
          "accounts": {
            "items": {
              "$ref": "#/components/schemas/ReportAccountResponse"
            },
            "title": "Accounts",
            "type": "array"
          },
          "close_sequence": {
            "title": "Close Sequence",
            "type": "integer"
          },
          "config_hash": {
            "title": "Config Hash",
            "type": "string"
          },
          "policy_hash": {
            "title": "Policy Hash",
            "type": "string"
          },
          "prices": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Prices",
            "type": "object"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "settlement_plan_hash": {
            "title": "Settlement Plan Hash",
            "type": "string"
          },
          "totals": {
            "$ref": "#/components/schemas/ReportTotalsResponse"
          },
          "trading_day": {
            "title": "Trading Day",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "trading_day",
          "close_sequence",
          "config_hash",
          "policy_hash",
          "settlement_plan_hash",
          "prices",
          "accounts",
          "totals"
        ],
        "title": "DailyReportResponse",
        "type": "object"
      },
      "DayInput": {
        "additionalProperties": false,
        "properties": {
          "trading_day": {
            "title": "Trading Day",
            "type": "string"
          }
        },
        "required": [
          "trading_day"
        ],
        "title": "DayInput",
        "type": "object"
      },
      "DayResponse": {
        "properties": {
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "phase": {
            "title": "Phase",
            "type": "string"
          },
          "trading_day": {
            "title": "Trading Day",
            "type": "string"
          }
        },
        "required": [
          "trading_day",
          "phase",
          "duplicate"
        ],
        "title": "DayResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReportAccountResponse": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "initial_capital": {
            "title": "Initial Capital",
            "type": "string"
          },
          "risk": {
            "$ref": "#/components/schemas/ReportRiskResponse"
          },
          "risk_event_count": {
            "title": "Risk Event Count",
            "type": "integer"
          },
          "settlement": {
            "$ref": "#/components/schemas/SettlementResultResponse"
          },
          "valuation": {
            "$ref": "#/components/schemas/ReportValuationResponse"
          }
        },
        "required": [
          "account_id",
          "initial_capital",
          "settlement",
          "valuation",
          "risk",
          "risk_event_count"
        ],
        "title": "ReportAccountResponse",
        "type": "object"
      },
      "ReportPositionResponse": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "floating_pnl": {
            "title": "Floating Pnl",
            "type": "string"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "string"
          },
          "margin": {
            "title": "Margin",
            "type": "string"
          },
          "notional": {
            "title": "Notional",
            "type": "string"
          },
          "product_id": {
            "title": "Product Id",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "side": {
            "$ref": "#/components/schemas/Side"
          }
        },
        "required": [
          "account_id",
          "instrument_id",
          "product_id",
          "side",
          "quantity",
          "notional",
          "margin",
          "floating_pnl"
        ],
        "title": "ReportPositionResponse",
        "type": "object"
      },
      "ReportRiskResponse": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "circuit_broken": {
            "title": "Circuit Broken",
            "type": "boolean"
          },
          "restricted_products": {
            "items": {
              "type": "string"
            },
            "title": "Restricted Products",
            "type": "array"
          },
          "warning": {
            "title": "Warning",
            "type": "boolean"
          }
        },
        "required": [
          "account_id",
          "warning",
          "circuit_broken",
          "restricted_products"
        ],
        "title": "ReportRiskResponse",
        "type": "object"
      },
      "ReportTotalsResponse": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "string"
          },
          "closing_balance": {
            "title": "Closing Balance",
            "type": "string"
          },
          "fees": {
            "title": "Fees",
            "type": "string"
          },
          "gross_exposure": {
            "title": "Gross Exposure",
            "type": "string"
          },
          "holding_pnl": {
            "title": "Holding Pnl",
            "type": "string"
          },
          "margin": {
            "title": "Margin",
            "type": "string"
          },
          "net_exposure": {
            "title": "Net Exposure",
            "type": "string"
          },
          "net_pnl": {
            "title": "Net Pnl",
            "type": "string"
          },
          "opening_balance": {
            "title": "Opening Balance",
            "type": "string"
          },
          "realized_pnl": {
            "title": "Realized Pnl",
            "type": "string"
          }
        },
        "required": [
          "opening_balance",
          "realized_pnl",
          "holding_pnl",
          "fees",
          "net_pnl",
          "closing_balance",
          "margin",
          "available",
          "gross_exposure",
          "net_exposure"
        ],
        "title": "ReportTotalsResponse",
        "type": "object"
      },
      "ReportValuationResponse": {
        "properties": {
          "available": {
            "title": "Available",
            "type": "string"
          },
          "equity": {
            "title": "Equity",
            "type": "string"
          },
          "floating_pnl": {
            "title": "Floating Pnl",
            "type": "string"
          },
          "gross_exposure": {
            "title": "Gross Exposure",
            "type": "string"
          },
          "margin": {
            "title": "Margin",
            "type": "string"
          },
          "net_exposure": {
            "title": "Net Exposure",
            "type": "string"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/ReportPositionResponse"
            },
            "title": "Positions",
            "type": "array"
          }
        },
        "required": [
          "positions",
          "floating_pnl",
          "equity",
          "margin",
          "available",
          "gross_exposure",
          "net_exposure"
        ],
        "title": "ReportValuationResponse",
        "type": "object"
      },
...
      "SettleInput": {
        "additionalProperties": false,
        "properties": {
          "prices": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Prices",
            "type": "object"
          },
          "trading_day": {
            "title": "Trading Day",
            "type": "string"
          }
        },
        "required": [
          "trading_day",
          "prices"
        ],
        "title": "SettleInput",
        "type": "object"
      },
...
      "SettlementLineResponse": {
        "properties": {
          "basis": {
            "title": "Basis",
            "type": "string"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "string"
          },
          "lot_id": {
            "title": "Lot Id",
            "type": "string"
          },
          "pnl": {
            "title": "Pnl",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "settlement_price": {
            "title": "Settlement Price",
            "type": "string"
          },
          "side": {
            "$ref": "#/components/schemas/Side"
          }
        },
        "required": [
          "lot_id",
          "instrument_id",
          "side",
          "quantity",
          "basis",
          "settlement_price",
          "pnl"
        ],
        "title": "SettlementLineResponse",
        "type": "object"
      },
      "SettlementResultResponse": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "available": {
            "title": "Available",
            "type": "string"
          },
          "closing_balance": {
            "title": "Closing Balance",
            "type": "string"
          },
          "fees": {
            "title": "Fees",
            "type": "string"
          },
          "holding_pnl": {
            "title": "Holding Pnl",
            "type": "string"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/SettlementLineResponse"
            },
            "title": "Lines",
            "type": "array"
          },
          "margin": {
            "title": "Margin",
            "type": "string"
          },
          "net_pnl": {
            "title": "Net Pnl",
            "type": "string"
          },
          "opening_balance": {
            "title": "Opening Balance",
            "type": "string"
          },
          "prices": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array"
            },
            "title": "Prices",
            "type": "array"
          },
          "realized_pnl": {
            "title": "Realized Pnl",
            "type": "string"
          },
          "trading_day": {
            "title": "Trading Day",
            "type": "string"
          }
        },
        "required": [
          "opening_balance",
          "realized_pnl",
          "holding_pnl",
          "fees",
          "net_pnl",
          "closing_balance",
          "margin",
          "available",
          "account_id",
          "trading_day",
          "prices",
          "lines"
        ],
        "title": "SettlementResultResponse",
        "type": "object"
      },
      "Side": {
        "enum": [
          "LONG",
          "SHORT"
        ],
        "title": "Side",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "CTA 期货风控 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/reports/{trading_day}": {
      "get": {
        "operationId": "getDailyReport",
        "parameters": [
          {
            "in": "path",
            "name": "trading_day",
            "required": true,
            "schema": {
              "title": "Trading Day",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report",
        "tags": [
          "settlement"
        ]
      }
    },
    "/api/reports/{trading_day}/download/{format}": {
      "get": {
        "operationId": "downloadDailyReport",
        "parameters": [
          {
            "in": "path",
            "name": "trading_day",
            "required": true,
            "schema": {
              "title": "Trading Day",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "format",
            "required": true,
            "schema": {
              "enum": [
                "html",
                "json"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download",
        "tags": [
          "settlement"
        ]
      }
    },
    "/api/reports/{trading_day}/export": {
      "post": {
        "operationId": "exportDailyReport",
        "parameters": [
          {
            "in": "path",
            "name": "trading_day",
            "required": true,
            "schema": {
              "title": "Trading Day",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Exportdailyreport",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export",
        "tags": [
          "settlement"
        ]
      }
    },
...
    "/api/settlement/close": {
      "post": {
        "operationId": "closeTradingDay",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DayInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Close Day",
        "tags": [
          "settlement"
        ]
      }
    },
    "/api/settlement/open": {
      "post": {
        "operationId": "openNextTradingDay",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DayInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Open Day",
        "tags": [
          "settlement"
        ]
      }
    },
    "/api/settlement/settle": {
      "post": {
        "operationId": "settleTradingDay",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettleInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Settle",
        "tags": [
          "settlement"
        ]
      }
    },
...
  }
}